VlasovLearn/requirements.txt
aioredis==1.3.1
amqp==2.6.1
asgi-redis==1.4.3
asgiref==3.3.1
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==4.4.7
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
defusedxml==0.7.0rc1
Django==2.2.17
django-google-analytics==2
django-google-analytics-app==5.0.2
django-google-charts==0.1.1
django-qsstats-magic==1.1.0
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==3.3.0
incremental==17.5.0
iniconfig==1.1.1
install==1.3.4
kombu==4.6.11
msgpack==1.0.2
msgpack-python==0.5.6
mypy==0.790
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.8
Pillow==8.0.1
pip==20.3.3
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.0.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypiwin32==223
pytest==6.2.1
python3-openid==3.2.0
python-dateutil==2.8.1
pytz==2020.5
pywin32==300
redis==2.10.6
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
setuptools==47.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.1
sqlparse==0.4.1
structlog==20.2.0
toml==0.10.2
Twisted==20.3.0
txaio==20.12.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==1.3.0
zipp==3.4.0
zope.interface==5.2.0
